+++ install.sh
@@ -2,7 +2,7 @@
 
 set -u
 
-version=v0.13
+version=v0.14
 
 APP=maple